+++ forge/src/main/java/net/zelythia/forge/ControllableCompatImpl.java
@@ -1,0 +1,7 @@
+package net.zelythia.forge;
+
+public class ControllableCompatImpl {
+    public static boolean attackDown() {
+        return false;
+    }
+}